Split views.py and its tests into four files each.
