+++ groups/tests/test_urls.py
@@ -8,14 +8,14 @@ class TestGroupUrls(URLTestCase):
 
     def test_group_list(self):
         self.assert_url_matches_view(
-            views.GroupList,
+            views.groups.GroupList,
             '/groups/',
             'group-list',
         )
 
     def test_group_detail(self):
         self.assert_url_matches_view(
-            views.GroupDetail,
+            views.groups.GroupDetail,
             '/groups/{}/'.format(self.pk),
             'group-detail',
             url_kwargs={'pk': self.pk}
@@ -23,7 +23,7 @@ class TestGroupUrls(URLTestCase):
 
     def test_group_subscribe(self):
         self.assert_url_matches_view(
-            views.GroupSubscribe,
+            views.groups.GroupSubscribe,
             '/groups/{}/subscribe/'.format(self.pk),
             'group-subscribe',
             url_kwargs={'pk': self.pk}
@@ -31,7 +31,7 @@ class TestGroupUrls(URLTestCase):
 
     def test_discussion_create(self):
         self.assert_url_matches_view(
-            views.DiscussionCreate,
+            views.discussions.DiscussionCreate,
             '/groups/{}/new-discussion/'.format(self.pk),
             'discussion-create',
             url_kwargs={'pk': self.pk}
@@ -39,7 +39,7 @@ class TestGroupUrls(URLTestCase):
 
     def test_discussion_detail(self):
         self.assert_url_matches_view(
-            views.DiscussionThread,
+            views.discussions.DiscussionThread,
             '/groups/discussions/{}/'.format(self.pk),
             'discussion-thread',
             url_kwargs={'pk': self.pk}
@@ -47,7 +47,7 @@ class TestGroupUrls(URLTestCase):
 
     def test_discussion_subscribe(self):
         self.assert_url_matches_view(
-            views.DiscussionSubscribe,
+            views.discussions.DiscussionSubscribe,
             '/groups/discussions/{}/subscribe/'.format(self.pk),
             'discussion-subscribe',
             url_kwargs={'pk': self.pk}
@@ -55,7 +55,7 @@ class TestGroupUrls(URLTestCase):
 
     def test_comment_upload_file(self):
         self.assert_url_matches_view(
-            views.CommentUploadFile,
+            views.comments.CommentUploadFile,
             '/groups/discussions/{}/upload-file/'.format(self.pk),
             'comment-upload-file',
             url_kwargs={'pk': self.pk}
@@ -63,7 +63,7 @@ class TestGroupUrls(URLTestCase):
 
     def test_comment_delete(self):
         self.assert_url_matches_view(
-            views.CommentDelete,
+            views.comments.CommentDelete,
             '/groups/comments/{}/delete/'.format(self.pk),
             'comment-delete',
             url_kwargs={'pk': self.pk}
