+++ groups/urls.py
@@ -4,51 +4,51 @@ from . import views
 
 
 urlpatterns = [
-    url(r'^$', views.GroupList.as_view(), name='group-list'),
+    url(r'^$', views.groups.GroupList.as_view(), name='group-list'),
     url(r'^(?P<pk>\d+)/', include([
         url(
             r'^$',
-            views.GroupDetail.as_view(),
+            views.groups.GroupDetail.as_view(),
             name='group-detail',
         ),
         url(
             r'^new-discussion/$',
-            views.DiscussionCreate.as_view(),
+            views.discussions.DiscussionCreate.as_view(),
             name='discussion-create',
         ),
         url(
             r'^subscribe/$',
-            views.GroupSubscribe.as_view(),
+            views.groups.GroupSubscribe.as_view(),
             name='group-subscribe',
         ),
     ])),
     url(r'^discussions/(?P<pk>\d+)/', include([
         url(
             r'^$',
-            views.DiscussionThread.as_view(),
+            views.discussions.DiscussionThread.as_view(),
             name='discussion-thread',
         ),
         url(
             r'^subscribe/$',
-            views.DiscussionSubscribe.as_view(),
+            views.discussions.DiscussionSubscribe.as_view(),
             name='discussion-subscribe',
         ),
         url(
             r'^upload-file/$',
-            views.CommentUploadFile.as_view(),
+            views.comments.CommentUploadFile.as_view(),
             name='comment-upload-file',
         ),
     ])),
     url(r'^comments/(?P<pk>\d+)/', include([
         url(
             r'^delete/$',
-            views.CommentDelete.as_view(),
+            views.comments.CommentDelete.as_view(),
             name='comment-delete',
         ),
     ])),
     url(
         r'^reply/',
-        views.CommentPostByEmail.as_view(),
+        views.comments.CommentPostByEmail.as_view(),
         name='comment-reply',
     ),
 ]
